code cleanup, refactored downsampler
